intelligence/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Time series analysis
statsmodels>=0.14.0
prophet>=1.1.0
pmdarima>=2.0.0

# Anomaly detection
pyod>=1.1.0
isolation-forest>=0.1.0

# NLP for query generation
spacy>=3.6.0
transformers>=4.30.0
nltk>=3.8.0

# Visualization analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0

# gRPC support
grpcio>=1.56.0
grpcio-tools>=1.56.0
protobuf>=4.23.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
ruff>=0.0.280
mypy>=1.4.0
ipykernel>=6.25.0